tracing-config/pyproject.toml
[project]
name = "tracing-config"
version = "0.1.0"
description = "OpenTelemetry tracing configuration utilities for the Self-Service Agent system"
authors = [
    { name = "Fabio Massimo Ercoli", email = "fabiomassimo.ercoli@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "opentelemetry-api>=1.37.0",
    "opentelemetry-sdk>=1.37.0",
    "protobuf>=6.33.5",
    "opentelemetry-exporter-otlp-proto-http>=1.37.0",
    "opentelemetry-instrumentation-httpx>=0.58b0",
]

[project.scripts]
tracing-config = "tracing_config:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/tracing_config"]

[tool.black]
line-length = 88
target-version = ['py312']

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.12"

[[tool.mypy.overrides]]
module = [
    "opentelemetry.*",
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tracing_config.*"
disable_error_code = ["attr-defined"]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=6.2.1",
    "black>=25.1.0",
    "flake8>=7.3.0",
    "isort>=5.13.0",
    "mypy>=1.17.1",
]
